agent/requirements.txt
langgraph
langchain-community
langchain-openai
tavily-python
tiktoken